software: more tweaks and optimisation
map: keep co-ordinates as floats for longer
	- goes faster on x86_64, might go slower on x86 and other.
pixmisc: sse2+ version use prefetching agian, mostly to add non-temporal hint

+++ src/software/palletblit.c
@@ -25,8 +25,20 @@
 
 //TODO: load pallets from files of some sort
 
+//TODO: optimize 16 bit modes
+//  - move y*w calculations outside of the loop
+//  - non-temporal store version if sse is available
+
+//TODO: dispatch based on availability of instructions
+//   except maybe on x86_64 since sse2 is always available there
+
 
 #if defined(__SSE2__)
+//TODO: if we have sse4.1 and dest and dst_stride are aligned properly
+// use non-temporal loads?
+
+//TODO: test if going up to a full 16 pixel at a time loop is faster, could be due to
+// that being a cache line size chuck
 static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, const uint16_t *restrict src, unsigned int w, unsigned int h, const uint32_t *restrict pal)
 {
 	// TODO: _mm_loadl_epi64 is dumb, it takes a pointer to __m128i rather than a uint64_t or __m64, figure out a way around using it.
@@ -46,8 +58,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 		switch(4 - ((dp%16)/4)) {
 			__m128i col1, col2, v1, v1s;
 			case 2:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, col2);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -68,8 +80,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				x+=2;
 			break;
 			case 3:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, col2);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -91,7 +103,7 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				x+=2;
 			//fall through
 			case 1:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, zero);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -120,8 +132,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 		for(; x + 8 < w; x+=8, s+=8, d+=8) {
 			__m128i col1, col2, col3, col4, v1, v2, v1s, v2s;
 
-			col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-			col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+			col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+			col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 			col1 = _mm_unpacklo_epi32(col1, col2);
 			col2 = col1;
 			col1 = _mm_unpacklo_epi8(col1, zero);
@@ -138,8 +150,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 			col1 = _mm_packus_epi16(col1, zero);
 
 
-			col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256)));
-			col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256)));
+			col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256u)));
+			col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256u)));
 
 			col3 = _mm_unpacklo_epi32(col3, col4);
 			col4 = col3;
@@ -159,8 +171,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 			_mm_stream_si128((__m128i *restrict)(d + 0), _mm_unpacklo_epi64(col1, col3));
 			//_mm_storeu_si128((__m128i *restrict)(d + 0), _mm_unpacklo_epi64(col1, col3));
 
-			col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[4]/256)));
-			col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[5]/256)));
+			col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[4]/256u)));
+			col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[5]/256u)));
 			col1 = _mm_unpacklo_epi32(col1, col2);
 			col2 = col1;
 			col1 = _mm_unpacklo_epi8(col1, zero);
@@ -177,8 +189,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 			col1 = _mm_packus_epi16(col1, zero);
 
 
-			col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[6]/256)));
-			col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[7]/256)));
+			col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[6]/256u)));
+			col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[7]/256u)));
 
 			col3 = _mm_unpacklo_epi32(col3, col4);
 			col4 = col3;
@@ -205,8 +217,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 		// pixels to force alignment and the loop does 8 pixels per iteration
 		// we end up with 4 - 8 pixels that need doing at loop exit)
 		__m128i col1, col2, col3, col4, v1, v2, v1s, v2s;
-		col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-		col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+		col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+		col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 		col1 = _mm_unpacklo_epi32(col1, col2);
 		col2 = col1;
 		col1 = _mm_unpacklo_epi8(col1, zero);
@@ -223,8 +235,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 		col1 = _mm_packus_epi16(col1, zero);
 
 
-		col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256)));
-		col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256)));
+		col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256u)));
+		col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256u)));
 
 		col3 = _mm_unpacklo_epi32(col3, col4);
 		col4 = col3;
@@ -248,8 +260,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 
 		switch(w-x) {
 			case 4:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, col2);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -266,8 +278,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				col1 = _mm_packus_epi16(col1, zero);
 
 
-				col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256)));
-				col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256)));
+				col3 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[2]/256u)));
+				col4 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[3]/256u)));
 
 				col3 = _mm_unpacklo_epi32(col3, col4);
 				col4 = col3;
@@ -287,8 +299,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				_mm_stream_si128((__m128i *restrict)(d + 0), _mm_unpacklo_epi64(col1, col3));
 			break;
 			case 2:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, col2);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -306,8 +318,8 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				_mm_stream_si64((int64_t *restrict)d, _mm_cvtsi128_si64(col1));
 			break;
 			case 3:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
-				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
+				col2 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[1]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, col2);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -328,7 +340,7 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 				d+=2;
 			//fall through
 			case 1:
-				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256)));
+				col1 = _mm_loadl_epi64((const __m128i *restrict)(pal+(s[0]/256u)));
 				col1 = _mm_unpacklo_epi32(col1, zero);
 				col2 = col1;
 				col1 = _mm_unpacklo_epi8(col1, zero);
@@ -365,9 +377,9 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 		const uint16_t *restrict s = src + y*w;
 		uint32_t *restrict d = (uint32_t *restrict)((char *restrict)dest + y*dst_stride);
 		for(size_t x = 0; x < w; x+=4, s+=4, d+=4) {
-			int v = s[0];
+			uint16_t v = s[0];
 
-			__m64 col1 = *(const __m64 *)(pal+(v/256));
+			__m64 col1 = *(const __m64 *)(pal+(v/256u));
 			__m64 col2 = col1;
 			col1 = _mm_unpacklo_pi8(col1, zero);
 			col2 = _mm_unpackhi_pi8(col2, zero);
@@ -384,7 +396,7 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 			__m64 tmp = col1;
 
 			v = s[1];
-			col1 = *(const __m64 *)(pal+(v/256));
+			col1 = *(const __m64 *)(pal+(v/256u));
 			col2 = col1;
 			col1 = _mm_unpacklo_pi8(col1, zero);
 			col2 = _mm_unpackhi_pi8(col2, zero);
@@ -405,7 +417,7 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 #endif
 
 			v = s[2];
-			col1 = *(const __m64 *)(pal+(v/256));
+			col1 = *(const __m64 *)(pal+(v/256u));
 			col2 = col1;
 			col1 = _mm_unpacklo_pi8(col1, zero);
 			col2 = _mm_unpackhi_pi8(col2, zero);
@@ -421,7 +433,7 @@ static void pallet_blit32(uint32_t *restrict dest, unsigned int dst_stride, cons
 			tmp = col1;
 
 			v = s[3];
-			col1 = *(const __m64 *)(pal+(v/256));
+			col1 = *(const __m64 *)(pal+(v/256u));
 			col2 = col1;
 			col1 = _mm_unpacklo_pi8(col1, zero);
 			col2 = _mm_unpackhi_pi8(col2, zero);
@@ -487,14 +499,14 @@ static void pallet_blit565(uint8_t *restrict dest, unsigned int dst_stride, cons
 		for(unsigned int x = 0; x < w; x+=4) {
 			__m64 r1, r2, g1, g2, b1, b2, c;
 			int v = src[y*src_stride + x];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r1 = _mm_shuffle_pi16(c, 0xfd);
 			g1 = _mm_shuffle_pi16(c, 0xfc);
 			c  = _mm_slli_pi32(c, 8);
 			b1 = _mm_shuffle_pi16(c, 0xfc);
 
 			v=src[y*src_stride + x+1];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0xf7);
 			g2 = _mm_shuffle_pi16(c, 0xf3);
 			c  = _mm_slli_pi32(c, 8);
@@ -504,7 +516,7 @@ static void pallet_blit565(uint8_t *restrict dest, unsigned int dst_stride, cons
 			b1 = _mm_or_si64(b1, b2);
 
 			v=src[y*src_stride + x+2];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0xdf);
 			g2 = _mm_shuffle_pi16(c, 0xcf);
 			c  = _mm_slli_pi32(c, 8);
@@ -514,7 +526,7 @@ static void pallet_blit565(uint8_t *restrict dest, unsigned int dst_stride, cons
 			b1 = _mm_or_si64(b1, b2);
 
 			v=src[y*src_stride + x+3];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0x7f);
 			g2 = _mm_shuffle_pi16(c, 0x3f);
 			c  = _mm_slli_pi32(c, 8);
@@ -548,14 +560,14 @@ static void pallet_blit555(uint8_t *restrict dest, unsigned int dst_stride, cons
 			__m64 r1, r2, g1, g2, b1, b2, c;
 
 			v  = src[y*src_stride + x];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r1 = _mm_shuffle_pi16(c, 0xfd);
 			g1 = _mm_shuffle_pi16(c, 0xfc);
 			c  = _mm_slli_pi32(c, 8);
 			b1 = _mm_shuffle_pi16(c, 0xfc);
 
 			v  = src[y*src_stride + x+1];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0xf7);
 			g2 = _mm_shuffle_pi16(c, 0xf3);
 			c  = _mm_slli_pi32(c, 8);
@@ -565,7 +577,7 @@ static void pallet_blit555(uint8_t *restrict dest, unsigned int dst_stride, cons
 			b1 = _mm_or_si64(b1, b2);
 
 			v  = src[y*src_stride + x+2];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0xdf);
 			g2 = _mm_shuffle_pi16(c, 0xcf);
 			c  = _mm_slli_pi32(c, 8);
@@ -575,7 +587,7 @@ static void pallet_blit555(uint8_t *restrict dest, unsigned int dst_stride, cons
 			b1 = _mm_or_si64(b1, b2);
 
 			v  = src[y*src_stride + x+3];
-			c  = _mm_cvtsi32_si64(pal[v/256]);
+			c  = _mm_cvtsi32_si64(pal[v/256u]);
 			r2 = _mm_shuffle_pi16(c, 0x7f);
 			g2 = _mm_shuffle_pi16(c, 0x3f);
 			c  = _mm_slli_pi32(c, 8);
